fix: remove CruiseApplicationApiHooks file to match current architecture
