--- frontend/src/api/hooks/applications/CruiseApplicationsApiHooks.tsx
@@ -1,155 +1,0 @@
-import { keepPreviousData, useInfiniteQuery, useMutation, useSuspenseQuery } from '@tanstack/react-query';
-
-import { client } from '@/lib/api';
-import { customFetch } from '@/api/client/custom-fetch';
-import { ApplicationResponse } from '@/api/client/applications/models';
-import { CruiseApplicationDto } from '@/api/dto/applications/CruiseApplicationDto';
-import { CruiseDto } from '@/api/dto/applications/ApplicationCruiseDto';
-import { EvaluationDto } from '@/api/dto/applications/EvaluationDto';
-import { UserEffectDto } from '@/api/dto/applications/UserEffectDto';
-import type { ApplicationPersonResponse } from '@/api/generated/schemas';
-
-export function useCruiseApplicationsQuery() {
-  return useSuspenseQuery({
-    queryKey: ['cruiseApplications'],
-    queryFn: async () => {
-      return client.get('/api/CruiseApplications');
-    },
-    select: (res) => res.data as CruiseApplicationDto[],
-  });
-}
-
-type CruiseApplicationsPage = {
-  items: ApplicationResponse[];
-  nextCursor: string | null;
-};
-
-export type CruiseApplicationsFilter = {
-  number?: number[];
-  date?: string[];
-  status?: string[];
-  year?: number[];
-  cruiseManager?: string[];
-};
-
-export type CruiseApplicationsSort = {
-  sortBy: string;
-  descending: boolean;
-};
-
-const DEFAULT_SORT: CruiseApplicationsSort = { sortBy: 'number', descending: true };
-
-const CRUISE_APPLICATIONS_PAGE_SIZE = 20;
-
-// Axios's default array serialization uses number[]=1&number[]=2, but ASP.NET Core's
-// [FromQuery] List<T> binding only recognizes repeated bare keys (number=1&number=2).
-function serializeRepeatedParams(params: Record<string, unknown>): string {
-  const searchParams = new URLSearchParams();
-  for (const [key, value] of Object.entries(params)) {
-    if (value === undefined || value === null) continue;
-    if (Array.isArray(value)) {
-      value.forEach((item) => searchParams.append(key, String(item)));
-    } else {
-      searchParams.append(key, String(value));
-    }
-  }
-  return searchParams.toString();
-}
-
-async function fetchCruiseApplicationsPage({
-  pageParam,
-  filter,
-  sort,
-}: {
-  pageParam: string | null;
-  filter: CruiseApplicationsFilter;
-  sort: CruiseApplicationsSort;
-}): Promise<CruiseApplicationsPage> {
-  const queryString = serializeRepeatedParams({
-    cursor: pageParam ?? undefined,
-    pageSize: CRUISE_APPLICATIONS_PAGE_SIZE,
-    sortBy: sort.sortBy,
-    descending: sort.descending,
-    ...filter,
-  });
-  return customFetch<CruiseApplicationsPage>(`/v2/applications?${queryString}`);
-}
-
-export function useCruiseApplicationsInfiniteQuery(
-  filter: CruiseApplicationsFilter = {},
-  sort: CruiseApplicationsSort = DEFAULT_SORT
-) {
-  return useInfiniteQuery({
-    queryKey: ['cruiseApplications', 'infinite', filter, sort],
-    queryFn: ({ pageParam }) => fetchCruiseApplicationsPage({ pageParam, filter, sort }),
-    initialPageParam: null as string | null,
-    getNextPageParam: (lastPage) => lastPage.nextCursor ?? undefined,
-    // Keeps the previous filter's rows (and the currently open filter dropdown) on
-    // screen while a new filter's page loads, instead of suspending the whole table.
-    placeholderData: keepPreviousData,
-    throwOnError: true,
-  });
-}
-
-export function useCruiseApplicationManagersQuery() {
-  return useSuspenseQuery({
-    queryKey: ['cruiseApplications', 'managers'],
-    queryFn: () => customFetch<ApplicationPersonResponse[]>('/v2/applications/managers'),
-  });
-}
-
-export function useCruiseApplicationQuery(id: string) {
-  return useSuspenseQuery({
-    queryKey: ['cruiseApplications', id],
-    queryFn: async () => {
-      return client.get(`/api/CruiseApplications/${id}`);
-    },
-    select: (res) => res.data as CruiseApplicationDto,
-  });
-}
-
-export function useCruiseForCruiseApplicationQuery(id: string) {
-  return useSuspenseQuery({
-    queryKey: ['cruiseApplication', id, 'cruise'],
-    queryFn: async () => {
-      return client.get(`/api/CruiseApplications/${id}/cruise`);
-    },
-    select: (res) => res.data as CruiseDto,
-  });
-}
-
-export function useEvaluationQuery(applicationId: string) {
-  return useSuspenseQuery({
-    queryKey: ['CruiseApplications', applicationId, 'evaluation'],
-    queryFn: async () => {
-      return client.get(`/api/CruiseApplications/${applicationId}/evaluation`);
-    },
-    select: (res) => res.data as EvaluationDto,
-  });
-}
-
-export function useEffectsEvaluationsQuery() {
-  return useSuspenseQuery({
-    queryKey: ['effectsEvaluations'],
-    queryFn: async () => {
-      return client.get('/api/CruiseApplications/effectsEvaluations');
-    },
-    select: (res) => res.data as UserEffectDto[],
-  });
-}
-
-export function useAcceptApplicationMutation() {
-  return useMutation({
-    mutationFn: async (applicationId: string) => {
-      return client.patch(`/api/CruiseApplications/${applicationId}/answer?accept=true`);
-    },
-  });
-}
-
-export function useRejectApplicationMutation() {
-  return useMutation({
-    mutationFn: async (applicationId: string) => {
-      return client.patch(`/api/CruiseApplications/${applicationId}/answer?accept=false`);
-    },
-  });
-}
